refactor(skills): move skills commands and trust roots into tau-skills
