+++ crates/tau-access/src/trust_roots.rs
@@ -1,165 +1,1 @@
-use std::path::Path;
-
-use anyhow::{anyhow, bail, Context, Result};
-use serde::{Deserialize, Serialize};
-use tau_core::write_text_atomic;
-use tau_skills::TrustedKey;
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TrustedRootRecord {
-    pub id: String,
-    pub public_key: String,
-    #[serde(default)]
-    pub revoked: bool,
-    pub expires_unix: Option<u64>,
-    pub rotated_from: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(untagged)]
-enum TrustedRootFileFormat {
-    List(Vec<TrustedRootRecord>),
-    Wrapped { roots: Vec<TrustedRootRecord> },
-    Keys { keys: Vec<TrustedRootRecord> },
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub struct TrustMutationReport {
-    pub added: usize,
-    pub updated: usize,
-    pub revoked: usize,
-    pub rotated: usize,
-}
-
-pub fn parse_trusted_root_spec(raw: &str) -> Result<TrustedKey> {
-    let (id, public_key) = raw
-        .split_once('=')
-        .ok_or_else(|| anyhow!("invalid --skill-trust-root '{raw}', expected key_id=base64_key"))?;
-    let id = id.trim();
-    let public_key = public_key.trim();
-    if id.is_empty() || public_key.is_empty() {
-        bail!("invalid --skill-trust-root '{raw}', expected key_id=base64_key");
-    }
-    Ok(TrustedKey {
-        id: id.to_string(),
-        public_key: public_key.to_string(),
-    })
-}
-
-pub fn parse_trust_rotation_spec(raw: &str) -> Result<(String, TrustedKey)> {
-    let (old_id, new_spec) = raw.split_once(':').ok_or_else(|| {
-        anyhow!("invalid --skill-trust-rotate '{raw}', expected old_id:new_id=base64_key")
-    })?;
-    let old_id = old_id.trim();
-    if old_id.is_empty() {
-        bail!("invalid --skill-trust-rotate '{raw}', expected old_id:new_id=base64_key");
-    }
-    let new_key = parse_trusted_root_spec(new_spec)?;
-    Ok((old_id.to_string(), new_key))
-}
-
-pub fn load_trust_root_records(path: &Path) -> Result<Vec<TrustedRootRecord>> {
-    if !path.exists() {
-        return Ok(Vec::new());
-    }
-    let raw = std::fs::read_to_string(path)
-        .with_context(|| format!("failed to read {}", path.display()))?;
-    let parsed = serde_json::from_str::<TrustedRootFileFormat>(&raw)
-        .with_context(|| format!("failed to parse trusted root file {}", path.display()))?;
-
-    let records = match parsed {
-        TrustedRootFileFormat::List(items) => items,
-        TrustedRootFileFormat::Wrapped { roots } => roots,
-        TrustedRootFileFormat::Keys { keys } => keys,
-    };
-
-    Ok(records)
-}
-
-pub fn save_trust_root_records(path: &Path, records: &[TrustedRootRecord]) -> Result<()> {
-    if let Some(parent) = path.parent() {
-        if !parent.as_os_str().is_empty() {
-            std::fs::create_dir_all(parent)
-                .with_context(|| format!("failed to create {}", parent.display()))?;
-        }
-    }
-    let mut payload = serde_json::to_string_pretty(&TrustedRootFileFormat::Wrapped {
-        roots: records.to_vec(),
-    })
-    .context("failed to serialize trusted root records")?;
-    payload.push('\n');
-    write_text_atomic(path, &payload)
-        .with_context(|| format!("failed to write {}", path.display()))?;
-    Ok(())
-}
-
-pub fn apply_trust_root_mutation_specs(
-    records: &mut Vec<TrustedRootRecord>,
-    add_specs: &[String],
-    revoke_ids: &[String],
-    rotate_specs: &[String],
-) -> Result<TrustMutationReport> {
-    let mut report = TrustMutationReport::default();
-
-    for spec in add_specs {
-        let key = parse_trusted_root_spec(spec)?;
-        if let Some(existing) = records.iter_mut().find(|record| record.id == key.id) {
-            existing.public_key = key.public_key;
-            existing.revoked = false;
-            existing.rotated_from = None;
-            report.updated += 1;
-        } else {
-            records.push(TrustedRootRecord {
-                id: key.id,
-                public_key: key.public_key,
-                revoked: false,
-                expires_unix: None,
-                rotated_from: None,
-            });
-            report.added += 1;
-        }
-    }
-
-    for id in revoke_ids {
-        let id = id.trim();
-        if id.is_empty() {
-            continue;
-        }
-        let record = records
-            .iter_mut()
-            .find(|record| record.id == id)
-            .ok_or_else(|| anyhow!("cannot revoke unknown trust key id '{}'", id))?;
-        if !record.revoked {
-            record.revoked = true;
-            report.revoked += 1;
-        }
-    }
-
-    for spec in rotate_specs {
-        let (old_id, new_key) = parse_trust_rotation_spec(spec)?;
-        let old = records
-            .iter_mut()
-            .find(|record| record.id == old_id)
-            .ok_or_else(|| anyhow!("cannot rotate unknown trust key id '{}'", old_id))?;
-        old.revoked = true;
-
-        if let Some(existing_new) = records.iter_mut().find(|record| record.id == new_key.id) {
-            existing_new.public_key = new_key.public_key;
-            existing_new.revoked = false;
-            existing_new.rotated_from = Some(old_id.clone());
-            report.updated += 1;
-        } else {
-            records.push(TrustedRootRecord {
-                id: new_key.id,
-                public_key: new_key.public_key,
-                revoked: false,
-                expires_unix: None,
-                rotated_from: Some(old_id.clone()),
-            });
-            report.added += 1;
-        }
-        report.rotated += 1;
-    }
-
-    Ok(report)
-}
+pub use tau_skills::trust_roots::*;
